refactoring init
